+++ lib/tir/purity.ml
@@ -18,7 +18,7 @@ let impure_builtins = [
   "file_read_line"; "file_open"; "file_close";
   "file_read"; "file_write"; "file_write_line";
   (* Network / TLS *)
-  "tcp_connect"; "tcp_send_all"; "tcp_recv_all"; "tcp_recv_exact"; "tcp_close";
+  "tcp_connect"; "tcp_send_all"; "tcp_recv_all"; "tcp_recv_exact"; "tcp_close"; "tcp_local_port";
   "tls_connect"; "tls_read"; "tls_write"; "tls_close";
   (* Randomness / non-determinism *)
   "random_bytes"; "stdlib_random_bytes";